requirements.txt
# Core dependencies for Job Description Generator
pydantic>=2.0.0
google-genai>=0.2.0
aiofiles>=23.2.0
